[ez][flakybot] Move some tests around (#6660)
Should be pretty much no change , just moving tests around
