+++ torchci/test/flakyBotTests/flakyBotUtils.test.ts
@@ -1,9 +1,8 @@
-import * as singleDisableIssue from "lib/flakyBot/singleDisableIssue";
 import * as flakyBotUtils from "lib/flakyBot/utils";
 import { IssueData } from "lib/types";
 import nock from "nock";
 import * as flakyBot from "pages/api/flaky-tests/disable";
-import { deepCopy, handleScope } from "test/common";
+import { handleScope } from "test/common";
 import {
   flakyTestA,
   flakyTestAcrossJobA,
@@ -294,21 +293,6 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     );
   });
 
-  test("getIssueTitle: test suite in subclass should not have __main__", async () => {
-    expect(
-      singleDisableIssue.getIssueTitle(
-        "test_cool_op_cpu",
-        "jit.async.SpecialSuite"
-      )
-    ).toEqual("DISABLED test_cool_op_cpu (jit.async.SpecialSuite)");
-  });
-
-  test("getIssueTitle: test suite not in subclass should be prefixed with __main__", async () => {
-    expect(
-      singleDisableIssue.getIssueTitle("test_cool_op_cpu", "TestLinAlgCPU")
-    ).toEqual("DISABLED test_cool_op_cpu (__main__.TestLinAlgCPU)");
-  });
-
   test("getWorkflowJobNames: should zip the workflow and job names of a test", async () => {
     expect(flakyBotUtils.getWorkflowJobNames(flakyTestA)).toEqual([
       "trunk / win-cpu-vs-2019 / test",
@@ -386,24 +370,6 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     expectJobsToDisablePlatforms(["dynamo linux"], ["dynamo"]);
   });
 
-  test("getIssueBodyForFlakyTest: should contain Platforms line", async () => {
-    expect(singleDisableIssue.getIssueBodyForFlakyTest(flakyTestA)).toContain(
-      "Platforms: "
-    );
-  });
-
-  test("getIssueBodyForFlakyTest: should contain correct examples URL", async () => {
-    expect(singleDisableIssue.getIssueBodyForFlakyTest(flakyTestA)).toContain(
-      "https://hud.pytorch.org/flakytest?name=test_a&suite=suite_a&limit=100"
-    );
-  });
-
-  test("getIssueBodyForFlakyTest: should contain file info", async () => {
-    expect(singleDisableIssue.getIssueBodyForFlakyTest(flakyTestA)).toContain(
-      "Test file path: `file_a.py`"
-    );
-  });
-
   test("filterOutNonFlakyTest: should not contain any flaky tests", async () => {
     const disabledNonFlakyTests = [nonFlakyTestA, nonFlakyTestZ];
 
@@ -454,262 +420,5 @@ describe("Disable Flaky Test Bot Unit Tests", () => {
     ]);
     await helper([closedSmall, closedBig], closedBig, []);
     await helper([closedSmall, openSmall], openSmall, []);
-  });
-
-  test("parseBody: bodyWithoutPlatforms preserves new lines", async () => {
-    expect(
-      singleDisableIssue.parseBody(
-        "Platforms: win, rocm\r\nHello\r\n\nMy name is\n\n```a;dlskfja\ndklfj```"
-      )
-    ).toEqual({
-      platformsToSkip: ["rocm", "win"],
-      invalidPlatforms: [],
-      bodyWithoutPlatforms:
-        "\r\nHello\r\n\nMy name is\n\n```a;dlskfja\ndklfj```",
-    });
-    expect(singleDisableIssue.parseBody("Platforms: win, rocm\r\n")).toEqual({
-      platformsToSkip: ["rocm", "win"],
-      invalidPlatforms: [],
-      bodyWithoutPlatforms: "\r\n",
-    });
-  });
-
-  describe("updateExistingIssueForFlakyTest", () => {
-    const defaultIssue = utils.genIssueData({});
-    beforeEach(() => {});
-
-    afterEach(() => {
-      nock.cleanAll();
-      jest.restoreAllMocks();
-    });
-
-    test("open issue, contains platforms", async () => {
-      const test = deepCopy(flakyTestA);
-
-      const scope = nock("https://api.github.com");
-      scope
-        .post(
-          `/repos/pytorch/pytorch/issues/${defaultIssue.number}/comments`,
-          (body) => {
-            const comment = JSON.stringify(body.body);
-            expect(comment).toContain("Another case of trunk flakiness");
-            expect(comment).toContain("appears to contain");
-            expect(comment).toContain("Either the change didn't propogate");
-            expect(comment.includes("Reopening issue")).toBe(false);
-            return true;
-          }
-        )
-        .reply(200, {});
-
-      await singleDisableIssue.updateExistingIssueForFlakyTest(
-        octokit,
-        defaultIssue,
-        test
-      );
-
-      handleScope(scope);
-    });
-
-    test("closed issue, contains platforms", async () => {
-      const test = deepCopy(flakyTestA);
-      const issue: IssueData = { ...defaultIssue, state: "closed" };
-
-      const scope = nock("https://api.github.com");
-      scope
-        .post(
-          `/repos/pytorch/pytorch/issues/${issue.number}/comments`,
-          (body) => {
-            const comment = JSON.stringify(body.body);
-            expect(comment).toContain("Another case of trunk flakiness");
-            expect(comment).toContain("appears to contain");
-            expect(comment).toContain("Reopening issue");
-            expect(comment.includes("Either the change didn't propogate")).toBe(
-              false
-            );
-            return true;
-          }
-        )
-        .reply(200, {})
-        .patch(`/repos/pytorch/pytorch/issues/${issue.number}`, (body) => {
-          expect(body.state).toEqual("open");
-          expect(body.body).toBe(undefined);
-          return true;
-        })
-        .reply(200, {});
-
-      await singleDisableIssue.updateExistingIssueForFlakyTest(
-        octokit,
-        issue,
-        test
-      );
-
-      handleScope(scope);
-    });
-
-    test("open issue, does not contain platforms", async () => {
-      const test = deepCopy(flakyTestA);
-      const issue: IssueData = {
-        ...defaultIssue,
-        body: "Platforms: linux\nhello",
-      };
-
-      const scope = nock("https://api.github.com");
-      scope
-        .post(
-          `/repos/pytorch/pytorch/issues/${issue.number}/comments`,
-          (body) => {
-            const comment = JSON.stringify(body.body);
-            expect(comment).toContain("Another case of trunk flakiness");
-            expect(comment).toContain("Adding [win]");
-            expect(comment.includes("Reopening issue")).toBe(false);
-            expect(comment.includes("Either the change didn't propogate")).toBe(
-              false
-            );
-            return true;
-          }
-        )
-        .reply(200, {})
-        .patch(`/repos/pytorch/pytorch/issues/${issue.number}`, (body) => {
-          expect(body.state).toEqual("open");
-          expect(body.body).toContain("Platforms: linux, win");
-          expect(body.body).toContain("hello");
-          return true;
-        })
-        .reply(200, {});
-
-      await singleDisableIssue.updateExistingIssueForFlakyTest(
-        octokit,
-        issue,
-        test
-      );
-
-      handleScope(scope);
-    });
-
-    test("closed issue, does not contain platforms", async () => {
-      const test = deepCopy(flakyTestA);
-      const issue: IssueData = {
-        ...defaultIssue,
-        body: "Platforms: linux\nhello",
-        state: "closed",
-      };
-
-      const scope = nock("https://api.github.com");
-      scope
-        .post(
-          `/repos/pytorch/pytorch/issues/${issue.number}/comments`,
-          (body) => {
-            const comment = JSON.stringify(body.body);
-            expect(comment).toContain("Another case of trunk flakiness");
-            expect(comment).toContain("does not appear to contain");
-            expect(comment).toContain("Adding [win]");
-            expect(comment).toContain("Reopening issue");
-            expect(comment.includes("Either the change didn't propogate")).toBe(
-              false
-            );
-            return true;
-          }
-        )
-        .reply(200, {})
-        .patch(`/repos/pytorch/pytorch/issues/${issue.number}`, (body) => {
-          expect(body.body).toContain("Platforms: linux, win");
-          expect(body.body).toContain("hello");
-          expect(body.state).toEqual("open");
-          return true;
-        })
-        .reply(200, {});
-
-      await singleDisableIssue.updateExistingIssueForFlakyTest(
-        octokit,
-        issue,
-        test
-      );
-
-      handleScope(scope);
-    });
-
-    test("open issue, does not contain platforms, longer platforms lists", async () => {
-      const test = deepCopy(flakyTestA);
-      test.jobNames.push("rocm");
-      test.workflowNames.push("test");
-      const issue: IssueData = {
-        ...defaultIssue,
-        body: "Platforms: inductor, dynamo\nhello",
-        state: "closed",
-      };
-
-      const scope = nock("https://api.github.com");
-      scope
-        .post(
-          `/repos/pytorch/pytorch/issues/${issue.number}/comments`,
-          (body) => {
-            const comment = JSON.stringify(body.body);
-            expect(comment).toContain("Another case of trunk flakiness");
-            expect(comment).toContain("does not appear to contain");
-            expect(comment).toContain("Adding [rocm, win]");
-            expect(comment).toContain("Reopening issue");
-            expect(comment.includes("Either the change didn't propogate")).toBe(
-              false
-            );
-            return true;
-          }
-        )
-        .reply(200, {})
-        .patch(`/repos/pytorch/pytorch/issues/${issue.number}`, (body) => {
-          expect(body.body).toContain("Platforms: dynamo, inductor, rocm, win");
-          expect(body.body).toContain("hello");
-          expect(body.state).toEqual("open");
-          return true;
-        })
-        .reply(200, {});
-
-      await singleDisableIssue.updateExistingIssueForFlakyTest(
-        octokit,
-        issue,
-        test
-      );
-
-      handleScope(scope);
-    });
-
-    test("closed issue, contains platforms, longer platforms lists", async () => {
-      const test = deepCopy(flakyTestA);
-      const issue: IssueData = {
-        ...defaultIssue,
-        body: "Platforms: win\nhello",
-        state: "closed",
-      };
-
-      const scope = nock("https://api.github.com");
-      scope
-        .post(
-          `/repos/pytorch/pytorch/issues/${issue.number}/comments`,
-          (body) => {
-            const comment = JSON.stringify(body.body);
-            expect(comment).toContain("Another case of trunk flakiness");
-            expect(comment).toContain("appears to contain");
-            expect(comment).toContain("Reopening issue");
-            expect(comment.includes("Either the change didn't propogate")).toBe(
-              false
-            );
-            return true;
-          }
-        )
-        .reply(200, {})
-        .patch(`/repos/pytorch/pytorch/issues/${issue.number}`, (body) => {
-          expect(body.body).toBe(undefined);
-          expect(body.state).toEqual("open");
-          return true;
-        })
-        .reply(200, {});
-
-      await singleDisableIssue.updateExistingIssueForFlakyTest(
-        octokit,
-        issue,
-        test
-      );
-
-      handleScope(scope);
-    });
   });
 });
